+++ src/node.ts
@@ -1,8 +1,18 @@
-export = {
-  env: {
-    node: true,
+import type { Linter } from "eslint";
+import globals from "globals";
+
+const config: Linter.Config[] = [
+  {
+    files: ["**/*.ts", "**/*.tsx", "**/*.js", "**/*.jsx"],
+    languageOptions: {
+      globals: {
+        ...globals.node,
+      },
+    },
+    rules: {
+      "no-path-concat": "error",
+    },
   },
-  rules: {
-    "no-path-concat": 2,
-  },
-};
+];
+
+export default config;